Add simple middleware pattern implementation
Contrary to other popular implementations this is very simple and
dependency container independant, which allows to use it on instanced
objects rather than registered services.

The explicit implementation is 30ns slower than plain method, has zero
runtime allocations and does not rely on reflection.

This is way faster than MediatR, so could be used in performance
sensitive applications.